requirements.txt
click>=8.1.7
pyyaml>=6.0.1
rich>=13.7.0
openai>=1.12.0
python-dotenv>=1.0.0
jsonschema>=4.21.0
requests>=2.31.0
